Macro expansion scopes are inserted at the last character in their insertion range
Macro expansion scope ranges are in terms of characters, so the end of the
insertion range is one past the last character of the insertion range. If
the last character in the insertion range is also the last character in
the enclosing scope, we would build the scope tree incorrectly.

Adjust the position for scope creation to the last character in the
insertion range, not one past it. Fixes rdar://120559184.